+++ gentlebot/cogs/AGENTS.md
