+++ gentlebot/cogs/test_logging_cog.py
@@ -3,8 +3,8 @@ import logging
 
 import discord
 from discord.ext import commands
-from util import chan_name
-import bot_config as cfg
+from ..util import chan_name
+from .. import bot_config as cfg
 
 # Use the same logger as main.py so handlers are attached
 log = logging.getLogger("gentlebot")
