+++ gentlebot/cogs/version_cog.py
@@ -11,8 +11,8 @@ import discord
 from discord import app_commands
 from discord.ext import commands
 
-from version import VERSION
-from util import chan_name
+from ..version import VERSION
+from ..util import chan_name
 
 # Use a hierarchical logger so messages propagate to the main gentlebot logger
 log = logging.getLogger(f"gentlebot.{__name__}")
