+++ gentlebot/util.py
