+++ test_harness.py
@@ -3,8 +3,6 @@ import logging
 import os
 
 import discord
-import sys
-from pathlib import Path
 
 os.environ.setdefault("env", "TEST")
 os.environ.setdefault("HF_API_TOKEN", "dummy")
@@ -14,9 +12,7 @@ os.environ.pop("PG_USER", None)
 os.environ.pop("PG_PASSWORD", None)
 os.environ.pop("PG_DB", None)
 
-from bot_config import TOKEN
-
-sys.path.insert(0, str(Path(__file__).resolve().parent / "src"))
+from gentlebot.bot_config import TOKEN
 from gentlebot.__main__ import GentleBot
