+++ tests/test_message_archive_cog.py
@@ -5,8 +5,8 @@ import discord
 from discord.ext import commands
 import asyncpg
 
-from cogs.message_archive_cog import MessageArchiveCog
-from util import build_db_url
+from gentlebot.cogs.message_archive_cog import MessageArchiveCog
+from gentlebot.util import build_db_url
 
 
 class DummyPool:
