+++ tests/test_postgres_handler.py
@@ -1,7 +1,7 @@
 import asyncio
 import logging
 import asyncpg
-from postgres_handler import PostgresHandler
+from gentlebot.postgres_handler import PostgresHandler
 
 class DummyPool:
     async def close(self):
